Eliminate MessagePipeDispatcher::incoming_messages_.
Instead, ReadMessageImplNoLock calls directly into Node::GetMessage.  However,
to support determining if the given buffer is large enough to hold the message,
a GetMessageIf function is added to Node. This allows ReadMessageImplNoLock to
probe the ports::Port's message queue before deciding whether or not to pop the
message off the queue. Some template magic is added to enable use of lambda
functions, which helps keep the implementation of ReadMessageImplNoLock in one
place.