Improve the fix for issue 5
It is not enough to simply defer the ObserveProxyAck. We also can't mutate the
peer link. In other words, a proxy pointing to a proxy that observes its peer
to be a proxy has to essentially ignore the event and upon its own death tell
its peer to re-send its ObserveProxy event.