allow SendMessage on uninitialized ports
this queues outgoing messages sent to ports that haven't been
fully initialized. at queue time, any sent ports are switched
to a buffering state as if the message were really being sent.

once the sending port is fully initialized, all queued messages
(and the ports they're sending) are modified to reflect the
now-known peer information before being flushed to the delegate
for forwarding.

this makes it possible to more easily implement disconnected
MPDs, which in turn simplifies interprocess bootstrapping logic.