add support for disconnected MPDs
it's now possible to construct an MPD from an uninitialized port.
a disconnected MPD functions normally, but any outgoing messages
are queued at the ports layer until the port is initialized.

an MPD may not be transferred until it's connected. this probably
won't matter in practice since disconnected pipes should only be
used for interprocess bootstrapping.