back to token-based bootstrap
<meme>this is the 9th time i've rewritten process bootstrap and
i've become exceedingly efficient at it</meme>

it turns out we're not going to escape changing a bunch of
process startup code anyway, and so using a platform channel
to negotiate port connections doesn't make sense.

it's still useful to be able to synchronously create disconnected
pipe endpoints, but we negotiate their connections using arbitrary
token strings.

this also brings a bunch more stuff into the overlay because
it's going to have to change :(