fix subtle race in MPD connection
it was possible for a recently connected port to wake up
a waiter with READABLE before OnPortStatusChanged was
invoked due to port initialization.

in this case, ReadMessage could then be called immediately
and return WAIT because port_connected_ was still false.

this change ensures that a port always acts disconnected
until the first port status change.