shutdown Channel asynchronously
a delegate might wish to shut down a channel in
OnChannelMessage, but this is currently not safe
because it could lead to synchronous Channel
destruction which would in turn lead to a UAF in
OnReadComplete and possibly in the Channel platform
impl.

this makes shutdown always happen asynchronously
when invoked through the public ShutDown interface.

it's possible that this breaks ChannelWin shutdown,
but i can't test at the moment.