add support for mojo::embedder::CreateChannel
*sigh*

well, in addition to token-based process bootstrap,
here we have (once again) bootstrap over platform
channels as another option. this is used indirectly
by mojo::embedder::CreateChannel, which in turn is
used a lot in content. converting everything would
be painful and messy, and it turns out that chrome
just works out of the box with this implementation,
so... here it is.

mojo_runner will continue to use tokens since that's
already done.