Implements support for platform handle passing on Windows.
Instead of a broker approach w/ synchronous IPCs, this solution routes
messages with platform handles through the parent process (node). The
parent process calls DuplicateHandle to map handles to the respective
child processes. This is achieved with a RelayPortsMessage call on
NodeChannel. Care is taken to avoid platform handle leaks.