call Create*MessagePipe callbacks without an implied thread hop
mojo_runner does weird things on shutdown, like stop running
the main thread message loop, then block on all children closing,
then destroy the message loop.

this could lead to deadlock if a ChildProcessHost tries to Join
after launching but before the pipe is created, because the
pipe callback would never run.

so the async embedder APIs no longer tries to be clever about calling
back on the calling thread, and consumers are left to ensure
thread-safety on their own.

ChildProcessHost has a thread-safe pipe holder which is used to
ensure pipe closure and thus clean shutdown in the aforementioned
case.