add tracking IDs to all pipes
this adds a persistent unique ID to every pipe. each endpoint
shares the same ID as its peer and is carried along with
the endpoint even when transferred across pipes.

the IDs are only meant to be used for tracking events
above the ports layer and should not be relied upon for
non-development, non-debugging, non-testing purposes.

[side note: this also downgrades ports-layer logging to
DVLOG(2) and fixes a one-line (but major!) bug in
ReadMessage.]